refactor : move tests out of sapphire library
