+++ Tests/UnitTests/Sources/BasicsTest/ReshapeTest.cpp
@@ -1,65 +1,62 @@
-// Copyright (c) 2021, Jaewoo Kim
+// Copyright (c) 2021, Justin Kim
 
 // We are making my contributions/submissions to this project solely in our
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <Sapphire/Tests/Basics/TransposeTest.hpp>
+#include <BasicsTest/ReshapeTest.hpp>
 #include <Sapphire/compute/BasicOps.hpp>
 #include <Sapphire/compute/Initialize.hpp>
 #include <Sapphire/util/Shape.hpp>
 #include <Sapphire/tensor/TensorData.hpp>
 #include <Sapphire/util/CudaDevice.hpp>
 #include <Sapphire/util/ResourceManager.hpp>
-#include <Sapphire/Tests/TestUtil.hpp>
+#include <Sapphire/compute/IndexingOps.hpp>
+#include <TestUtil.hpp>
 #include <Sapphire/Model.hpp>
 #include <iostream>
 #include <random>
+#include "doctest.h"
 
 namespace Sapphire::Test
 {
-void TransposeTest(bool printResult)
+void ReshapeTest(bool printResult)
 {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<> distrib(1, 100);
+    std::uniform_int_distribution<> dist(1, 100);
 
-    const int dim = distrib(gen) % 1 + 2;
+    const int dim = dist(gen) % 1 + 2;
 
     const auto shapeInput = CreateRandomShape(dim, 30);
-    const auto shapeTransposed = shapeInput.GetTranspose();
+    const auto newShape = shapeInput.GetTranspose();
 
     const CudaDevice cuda(0, "device0");
 
     //! Define TensorData for input and transposed
     TensorUtil::TensorData inputTensor(shapeInput, Type::Dense, cuda);
-    TensorUtil::TensorData transposedTensor(shapeTransposed, Type::Dense, cuda);
     inputTensor.SetMode(ComputeMode::Host);
-    transposedTensor.SetMode(ComputeMode::Host);
 
     //! Initialize input Tensor with normal distribution
     Compute::Initialize::Normal(inputTensor, 10, 5);
     //! Initialize output Tensor with zeros
-    Compute::Initialize::Zeros(transposedTensor);
+    TensorUtil::TensorData transposedTensor = inputTensor.CreateCopy();
 
-    //! Perform transpose on the host side
-    Compute::Transpose(transposedTensor, inputTensor);
+    //! Perform reshape on host
+    transposedTensor.Reshape(newShape);
+    CHECK(transposedTensor.GetShape() == newShape);
+
     //! Use the buffer to store the host result temporarily
-    auto* cpuResult = new float[transposedTensor.HostTotalSize];
-    std::memcpy(cpuResult, transposedTensor.HostRawPtr(),
-                transposedTensor.HostTotalSize * sizeof(float));
-
-    //! Re-initialize the transoised tensor with zeros
-    Compute::Initialize::Zeros(transposedTensor);
+    auto cpuResult = transposedTensor.GetDataCopy();
 
     //! Send the input tensor to cuda
     inputTensor.ToCuda();
-    transposedTensor.ToCuda();
     inputTensor.SetMode(ComputeMode::Cuda);
-    transposedTensor.SetMode(ComputeMode::Cuda);
+    transposedTensor = inputTensor.CreateCopy();
 
-    //! Transpose the tensor on cuda
-    Compute::Transpose(transposedTensor, inputTensor);
+    //! Perform reshape on cuda
+    transposedTensor.Reshape(newShape);
+    CHECK(transposedTensor.GetShape() == newShape);
 
     //! Send the transposed result to host
     //! Zero initialized data on the host should be overwritten
@@ -67,10 +64,8 @@ void TransposeTest(bool printResult)
     transposedTensor.SetMode(ComputeMode::Host);
 
     //! Compare the results
-    const float* cudaResult = transposedTensor.HostRawPtr();
-    CheckNoneZeroEquality(cpuResult, cudaResult,
-                          transposedTensor.HostTotalSize, printResult);
-
-    delete[] cpuResult;
+    auto cudaResult = transposedTensor.GetDataCopy();
+    CheckNoneZeroEquality(std::move(cpuResult), std::move(cudaResult),
+                          transposedTensor.Size(), printResult);
 }
 }
